src/bad-brains-service/requirements.txt
fastapi==0.115.5
httpx==0.18.1
pydantic==1.8.2
pymongo==4.10.1
PyYAML==6.0.2
uvicorn==0.11.8
